requirements.txt
Flask
python-dotenv
mysql-connector-python